add custom multiline test
